fix: map font size loading
